Update SDK, configurations, properties, and launch settings
Updated the project SDK from `Microsoft.NET.Sdk.WindowsDesktop` to `Microsoft.NET.Sdk`. Added Debug and Release configurations for R25 and several new properties to the PropertyGroup. A new PropertyGroup was added for the 'R25' configuration with a target framework of `net8.0-windows` and a Revit version of 2025. Package references for `Nice3point.Revit.Api.RevitAPI`, `Nice3point.Revit.Api.RevitAPIUI`, and `Nice3point.Revit.Api.AdWindows` were replaced with a conditional statement. Added a new profile to the launch settings for Revit 2025 and removed the profile for Revit 2018.