Added sleep command to pause program execution
In this commit, a sleep command has been added to pause the execution of the program for 2 seconds. This is to ensure that the build process initiated by the previous command has enough time to complete before the program continues. The sleep command is implemented using `System.Threading.Thread.Sleep(2000);`, where 2000 represents the number of milliseconds to sleep.